+++ language/it/abbc3.php
@@ -76,12 +76,7 @@ $lang = array_merge($lang, array(
 	// BBCode Wizards
 	'ABBC3_BBCODE_WIZ_SUBMIT'	=> 'Inserire nel messaggio',
 	'ABBC3_BBCODE_WIZ_EXAMPLE'	=> 'Esempio',
-
-	// BBvideo Wizard
 	'ABBC3_BBVIDEO_SITES'		=> 'Siti abilitati per i video',
-	'ABBC3_BBVIDEO_LINK'		=> 'URL video',
-
-	// URL Wizard
 	'ABBC3_URL_LINK'			=> 'Aggiungi un URL',
 	'ABBC3_URL_DESCRIPTION'		=> 'Descrizione opzionale',
 	'ABBC3_URL_EXAMPLE'			=> 'http://www.phpbb.com',