+++ tests/controller/wizard_test.php
@@ -132,7 +132,6 @@ class wizard_test extends \phpbb_test_case
 			->method('assign_vars')
 			->with([
 				'ABBC3_BBVIDEO_SITES'   => [],
-				'ABBC3_BBVIDEO_LINK_EX' => '',
 				'ABBC3_BBVIDEO_DEFAULT' => \vse\abbc3\controller\wizard::BBVIDEO_DEFAULT,
 			]);