+++ tests/event/quick_reply_test.php
@@ -8,26 +8,31 @@
 *
 */
 
-namespace vse\abbc3\event;
+namespace vse\abbc3\tests\event;
 
 class quick_reply_test extends \vse\abbc3\tests\event\listener_base
 {
 	public function set_quick_reply_data()
 	{
 		return [
-			[1],
-			[0],
+			[1, 2],
+			[0, 2],
 		];
 	}
 
 	/**
 	 * @dataProvider set_quick_reply_data
 	 */
-	public function test_set_quick_reply($enabled)
+	public function test_set_quick_reply($enabled, $forum_id)
 	{
 		$this->set_listener();
 
 		$this->config['abbc3_qr_bbcodes'] = $enabled;
+
+		$this->bbcodes_display->expects($enabled ? self::once() : self::never())
+			->method('bbcode_statuses')
+			->with($forum_id)
+			->willReturn([]);
 
 		$this->user->expects($enabled ? self::once() : self::never())
 			->method('add_lang')
@@ -35,13 +40,9 @@ class quick_reply_test extends \vse\abbc3\tests\event\listener_base
 
 		$this->template->expects($enabled ? self::once() : self::never())
 			->method('assign_var')
-			->with('S_BBCODE_ALLOWED', true);
+			->with('S_ABBC3_QUICKREPLY', true);
 
 		$this->listener->set_quick_reply();
-		$this->listener->add_to_quickreply();
+		$this->listener->add_to_quickreply(['forum_id' => $forum_id]);
 	}
-}
-
-function display_custom_bbcodes()
-{
 }