+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.3.1 - 2020-11-07
+
+- Joe Biden wins the US Presidential election.
+- Fixed missing IMG, URL and FLASH BBCode buttons from Quick Reply.
+- Fixed issues with duplication of BBCode button bars in Quick Reply when using Quick Reply Reloaded and other Quick Reply extensions.
+- Fixed the names of example sites in the BBVideo/Media Embed Wizard.
+- Added European Portuguese translation
+
 ## 3.3.0 - 2020-09-11
 
 - Replaced the antiquated, pixelated GIF BBCode icons with two new SVG and PNG icon sets.